requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
aiosmtplib~=3.0.1
aioimaplib~=1.1.0
PyJWT~=2.8.0
pip~=24.0
attrs~=23.2.0
wheel~=0.43.0
pillow~=10.3.0
filelock~=3.15.4
Babel~=2.15.0
greenlet~=3.0.3
PyMySQL~=1.1.1
aiofiles==23.1.0
pysocks==1.7.1
Flask[async]==3.0.3
requests==2.32.3
telethon==1.36.0
